# swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Package main",
    "version": "0.0.1"
  },
  "host": "localhost:10000",
  "basePath": "/",
  "paths": {
...
    "/api/task/group": {
      "get": {
        "description": "Returns a list of all TaskGroups",
        "tags": [
          "TaskGroup"
        ],
        "operationId": "GetAllTaskGroupsEndpoint",
        "responses": {
          "200": {
            "description": "TaskGroups response",
            "schema": {
              "$ref": "#/responses/TaskGroupsResponseSwagger"
            }
          }
        }
      },
      "post": {
        "description": "Creates a TaskGroup in the database",
        "tags": [
          "TaskGroup"
        ],
        "operationId": "CreateTaskGroupEndpoint",
        "parameters": [
          {
            "description": "TaskGroup details",
            "name": "TaskGroupDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateTaskGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TaskGroup response",
            "schema": {
              "$ref": "#/responses/TaskGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/group/{GroupID}": {
      "get": {
        "tags": [
          "TaskGroup"
        ],
        "summary": "Returns the TaskGroups with GroupID {GroupID}.",
        "operationId": "GetTaskGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of TaskGroups to retrieve",
            "name": "GroupID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "TaskGroups response",
            "schema": {
              "$ref": "#/responses/TaskGroupsResponseSwagger"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TaskGroup"
        ],
        "summary": "Updates and returns the TaskGroup with GroupID {GroupID}.",
        "operationId": "UpdateTaskGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of TaskGroup to update",
            "name": "GroupID",
            "in": "path"
          },
          {
            "description": "Details to update TaskGroup with",
            "name": "TaskGroupDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateTaskGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TaskGroup response",
            "schema": {
              "$ref": "#/responses/TaskGroupResponseSwagger"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TaskGroup"
        ],
        "summary": "Deletes and returns the TaskGroup with GroupID {GroupID}.",
        "operationId": "DeleteTaskGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of TaskGroup to retrieve",
            "name": "GroupID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "TaskGroup response",
            "schema": {
              "$ref": "#/responses/TaskGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/group/{GroupID}/clone": {
      "post": {
        "tags": [
          "TaskGroup"
        ],
        "summary": "Clones the TaskGroup with GroupID {GroupID} and returns the clone.",
        "operationId": "CloneTaskGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of TaskGroup to clone",
            "name": "GroupID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "TaskGroup response",
            "schema": {
              "$ref": "#/responses/TaskGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/group/{GroupID}/removeTasks": {
      "post": {
        "tags": [
          "TaskGroup"
        ],
        "summary": "Deletes Tasks from the group with GroupID {GroupID}.",
        "operationId": "RemoveTasksEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "GroupID of Tasks to remove",
            "name": "GroupID",
            "in": "path",
            "required": true
          },
          {
            "description": "IDs of Tasks to remove from Group",
            "name": "TaskIDs",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/{GroupID}": {
      "post": {
        "description": "Creates a Task in the database",
        "tags": [
          "Task"
        ],
        "operationId": "CreateTaskEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of TaskGroup to add task to",
            "name": "GroupID",
            "in": "path"
          },
          {
            "description": "Task details",
            "name": "TaskDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/{ID}": {
      "get": {
        "tags": [
          "Task"
        ],
        "summary": "Returns the Task with ID {ID}.",
        "operationId": "GetTaskEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Task to retrieve",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Task"
        ],
        "summary": "Updates and returns the Task with ID {ID}.",
        "operationId": "UpdateTaskEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Task to update",
            "name": "ID",
            "in": "path"
          },
          {
            "description": "Details to update Task with",
            "name": "TaskDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateTaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/{ID}/clone": {
      "post": {
        "tags": [
          "Task"
        ],
        "summary": "Clones the Task with ID {ID} and returns the clone.",
        "operationId": "CloneTaskEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Task to clone",
            "name": "ID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/{ID}/start": {
      "post": {
        "description": "Starts a Task",
        "tags": [
          "Task"
        ],
        "operationId": "StartTaskEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Task to start",
            "name": "ID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      }
    },
    "/api/task/{ID}/stop": {
      "post": {
        "description": "Stops a Task",
        "tags": [
          "Task"
        ],
        "operationId": "StopTaskEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Task to stop",
            "name": "ID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Task response",
            "schema": {
              "$ref": "#/responses/TaskResponseSwagger"
            }
          }
        }
      }
    }
  }
}